+++ Experiments.h
@@ -12,10 +12,10 @@ public:
 	Experiments();
 	Experiments(const Experiments & exp);
 	Experiments removeMutations(Specie *s);
-	bool addTimePoint(int experiment, int row, int column, float data);
-	float getTimePoint(int experiment, int row, int column);
-	std::vector<float> getSortedValues(int column);
-	std::vector<float> * getRow(int experiment, int row);
+	bool addTimePoint(int experiment, int row, int column, double data);
+	double getTimePoint(int experiment, int row, int column);
+	std::vector<double> getSortedValues(int column);
+	std::vector<double> * getRow(int experiment, int row);
 	int totalExperiments();
 	int totalRows(int experiment);
 	virtual ~Experiments();
@@ -23,7 +23,7 @@ public:
 	const Experiments & operator=(const Experiments & e);
 
 private:
-	vector<vector<vector<float > *> *> * experiments;
+	vector<vector<vector<double > *> *> * experiments;
 	void deleteExperiments();
 	void fillExperiments(const Experiments & exp);
 };
